refactor: move authenticate form
